Added NRE for OfHoraxCult
When loading the game I got NREs for some stranger pawns from this method, like the VQEGenerator's Genetron.inventor.
After investigating it turned out that OfHoraxCult can result in NREs thrown by the game if the World and/or FactionManager have not finished initializing yet.
I don't know if this is really the most elegant way to solve it, it feels hacky, but at least it fixed the on load errors for me.

Example NRE:

Error while determining if Simi should have Need VAEI_Sanity: System.Reflection.TargetInvocationException: Exception has been thrown by the target of an invocation. ---> System.NullReferenceException: Object reference not set to an instance of an object